+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 CHANGELOG
 =========
 
+0.0.2 (2016-04-18)
+------------------
+* Renamed Position model to post because of naming clashes
+* Added is_public post
+* Renamed is_active flag to active_post
+* Removed is_public again
+* Added CKEditor Dependency
+* Using HTMLField for description
+* Added experimental templates
+
+
 0.0.1 (2016-04-18)
 ------------------